+++ libmvsim/src/VehicleDynamics/VehicleAckermann_ControllerTwistFrontSteerPID.cpp
@@ -1,10 +1,10 @@
 /*+-------------------------------------------------------------------------+
   |                       MultiVehicle simulator (libmvsim)                 |
   |                                                                         |
-  | Copyright (C) 2014  Jose Luis Blanco Claraco (University of Almeria)    |
+  | Copyright (C) 2014-2020  Jose Luis Blanco Claraco                       |
   | Copyright (C) 2017  Borys Tymchenko (Odessa Polytechnic University)     |
-  | Distributed under GNU General Public License version 3                  |
-  |   See <http://www.gnu.org/licenses/>                                    |
+  | Distributed under 3-clause BSD License                                  |
+  |   See COPYING                                                           |
   +-------------------------------------------------------------------------+ */
 
 #include <mvsim/VehicleDynamics/VehicleAckermann.h>